temp Final

+++ routes/api.php
@@ -12,27 +12,28 @@ use Illuminate\Support\Facades\Route;
 // });
 
 Route::post('/login', [AuthApiController::class, 'login'])->name('login');
-Route::post('/admin/logout', [AuthApiController::class, 'adminLogout'])->name('adminLogout');
+Route::post('/logout', [AuthApiController::class, 'teamLogout'])->name('logout')->middleware('isLogin');;
+Route::post('/admin/logout', [AuthApiController::class, 'adminLogout'])->name('adminLogout')->middleware('isAdmin');;;
 Route::post('/admin/login', [AuthApiController::class, 'adminLogin'])->name('adminLogin');
 Route::post('/register', [AuthApiController::class, 'register']);
 
 
-Route::get('/user/{id}', [UserController::class, 'getUserById']);
-Route::get('/team/{teamId}/users', [UserController::class, 'getUsersByTeamId']);
+Route::get('/user/{id}', [UserController::class, 'getUserById'])->middleware('isAdmin');;
+Route::get('/team/{teamId}/users', [UserController::class, 'getUsersByTeamId'])->middleware('isAdmin');;
 
-Route::post('/teams', [TeamController::class,'createTeam']);
+Route::post('/teams', [TeamController::class,'createTeam'])->middleware('isAdmin');;
 Route::get('/teams', [TeamController::class,'teamList'])->name('teamList')->middleware('isAdmin');
-Route::get('/teams/name/{stat}',[TeamController::class,'showTeamSortbyName'])->name('sortByName');
-Route::get('/teams/time/{stat}',[TeamController::class,'showTeamSortbyTime'])->name('sortByTime');
-Route::get('/teams/name/{name}',[TeamController::class,'searchByName']);
-Route::get('/teams/{teamId}/edit',[TeamController::class, 'editTeam'])->name('editTeam');
-Route::put('/teams/{teamId}', [TeamController::class, 'updateTeam']);
-Route::delete('/teams/{teamId}', [TeamController::class,'deleteTeam'])->name('deleteTeam');
+Route::get('/teams/name/{stat}',[TeamController::class,'showTeamSortbyName'])->name('sortByName')->middleware('isAdmin');;
+Route::get('/teams/time/{stat}',[TeamController::class,'showTeamSortbyTime'])->name('sortByTime')->middleware('isAdmin');;
+Route::get('/teams/name/{name}',[TeamController::class,'searchByName'])->name('searchByName')->middleware('isAdmin');;
+Route::get('/teams/{teamId}/edit',[TeamController::class, 'editTeam'])->name('editTeam')->middleware('isAdmin');
+Route::put('/teams/{teamId}', [TeamController::class, 'updateTeam'])->name('updateTeam')->middleware('isAdmin');
+Route::delete('/teams/{teamId}', [TeamController::class,'deleteTeam'])->name('deleteTeam')->middleware('isAdmin');
 
 // view controller
 
 Route::get('/', [ViewController::class, 'viewHome'])->name('viewHome');
 Route::get('/login', [ViewController::class, 'viewLogin'])->name('loginView');
-Route::get('/dashboard/{teamId}', [ViewController::class, 'viewUserDashboard'])->name('dashboardView');
+Route::get('/dashboard', [ViewController::class, 'viewUserDashboard'])->name('dashboardView')->middleware('isLogin');
 Route::get('/admin', [ViewController::class,"viewAdminLogin"])->name('viewAdminLogin');
 Route::get('/admin/dashboard', [TeamController::class,"teamList"])->name('adminDashboardView')->middleware('isAdmin');
